requirements.txt
openai>=1.40.0
langgraph>=0.2.32
pandas>=2.2.2
numpy>=1.26.4
yfinance>=0.2.52
ta>=0.11.0
streamlit>=1.37.0
python-dotenv>=1.0.1
pydantic>=2.8.2
duckduckgo-search>=5.3.2
plotly>=5.23.0
httpx>=0.27.0
tenacity>=8.5.0
